fix shared_library for windows
The shared_library path is problematic on Windows because ctypesgen
inserts the path straight into the final file, which ends up being
interpreted as escape characters.

This is surprising because it means that ctypesgen has probably never
been used to generate bindings on Windows. I don't see how this would
have ever worked in the past for anyone.

Temporary fix is to directly use the DLL filename and assume that brlcad
was installed "system-wide" (it's an option during brlcad installation,
but it's not a default...).

I think there are other problems with non-system-wide installations of
brlcad related to tcl.dll and some MSVCR-related DLL files, so this is
just an additional reason why system-wide brlcad installations on
Windows are mandatory.